server/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
geographiclib==1.52
geopy==2.2.0
idna==3.4
imageio==2.22.2
kiwisolver==1.4.4
matplotlib==3.6.0
networkx==2.8.7
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
svgwrite==1.4.3
tifffile==2022.10.10
torch==1.9.0
torchvision==0.10.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
